bind::zone: add ability for update-policy{}
Allows for matchtype of name and subdomain with optionally specifying
resource records.

Add update_policy parameter to bind::zone which, if specified, must be a
hash in the following format. Note that matchtype and key are mandatory
and rrs is an optional array.

bind::zones:
  'example.com':
    type: 'master'
    target: '/etc/named/zone_lists/internal.zones'
    tag: 'internal'
    extra_path: '/internal'
    update_policies:
      'dev.example.com':
        matchtype: 'subdomain'
        key: 'key-internal'
      'foo.example.com':
        matchtype: 'name'
        key: 'some-key'
        rrs:
          - 'CNAME'

would result in

zone "example.com" {
  type master;
  file "master/internal/example.com";

  update-policy
  {
    grant key-internal subdomain dev.example.com.;
    grant some-key name foo.example.com. CNAME;
  };
};